Created shareable GenreListItem and cleaned up Library Genres
